PCT-565 Protection from invalid caputure
Added a control of time took to collect the metrics.
If it is > 10% of ticker interval, metrics are discarded.
The calculated time include all metrics collected in the run function
because ie if SHOW STATUS takes 3 seconds, all following metrics like
SELECT NAME, ... FROM INFORMATION_SCHEMA.INNODB_METRICS will be
timeshifted, so I think we need to ignore them to keep them synchronized.